Refactoring: Move ArbitratorModule to parent package
